refactor(libs-state-entity): move actions & services to withConfig
refactor(libs-state-entity): move actions & services to withConfig
