+++ libs/shared/interface/entity/src/actions.ts
@@ -1,0 +1,5 @@
+export enum ActionsEntity {
+  ASSIGN_ITEM = 'ASSIGN_ITEM',
+  ASSIGN_LIST = 'ASSIGN_LIST',
+  FETCH_DATA = 'FETCH_DATA',
+}
