+++ libs/shared/interface/entity/src/index.ts
@@ -1,1 +1,3 @@
 export * from './entity'
+export * from './actions'
+export * from './services'
