+++ libs/shared/interface/entity/src/services.ts
@@ -1,0 +1,4 @@
+export enum ServicesEntity {
+  FETCH_LIST = 'FETCH_LIST',
+  FETCH_ITEM = 'FETCH_ITEM',
+}
