+++ libs/shared/interface/machine/src/index.ts
@@ -1,1 +1,1 @@
-export * from './options-machine'
+export * from './options'
