--- libs/shared/interface/machine/src/options-machine.ts
@@ -1,11 +1,0 @@
-import { EventObject, MachineOptions, ServiceConfig } from 'xstate'
-
-export interface CustomMachineOptions<
-  TContext = any,
-  TEvent extends EventObject = any
-> extends Partial<MachineOptions<TContext, TEvent>> {
-  services: {
-    fetchItem: ServiceConfig<TContext, TEvent>
-    fetchList: ServiceConfig<TContext, TEvent>
-  }
-}
